+++ src/app/page.tsx
@@ -1,498 +1,327 @@
 "use client";
 
-import React, { useState, useEffect, useCallback } from "react";
-import { createBrowserClient } from "@supabase/ssr";
-import { useRouter } from "next/navigation";
-import type { User } from "@supabase/supabase-js";
-import Timetable from "./components/Timetable";
-import RescheduleModal from "./components/RescheduleModal";
-import SettingsModal from "./components/SettingsModal";
-import { Calendar } from "./components/ui/calendar";
-import { Button } from "./components/ui/button";
-import { Slider } from "./components/ui/slider";
-import { PlusIcon } from "lucide-react";
+import React from "react";
+import Link from "next/link";
+import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
+import styles from "./marketing.module.css";
+import Image from "next/image";
 
-const PlazenLogo = () => (
-  <svg
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    className="text-white"
-  >
-    <path
-      d="M3 12C5.66667 8 7.33333 8 10 12C12.6667 16 14.3333 16 17 12C19.6667 8 21 8 21 8"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-    <circle cx="6" cy="10" r="1.5" fill="currentColor" />
-    <circle cx="12" cy="14" r="1.5" fill="currentColor" />
-    <circle cx="18" cy="10" r="1.5" fill="currentColor" />
-  </svg>
-);
+// Adjust these to match your auth and routes:
+const SESSION_COOKIE = "plazen_session";
+const SCHEDULE_PATH = "/schedule";
 
-const UserIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="h-6 w-6 text-gray-400"
-  >
-    <path d="M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2"></path>
-    <circle cx="12" cy="7" r="4"></circle>
-  </svg>
-);
+export default function Page() {
+  return <Landing />;
+}
 
-const SettingsIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="h-6 w-6 text-gray-400"
-  >
-    <path d="M12.22 2h-.44a2 2 0 0 0-2 2v.18a2 2 0 0 1-1 1.73l-.43.25a2 2 0 0 1-2 0l-.15-.08a2 2 0 0 0-2.73.73l-.22.38a2 2 0 0 0 .73 2.73l.15.1a2 2 0 0 1 0 2l-.15.08a2 2 0 0 0-.73 2.73l.22.38a2 2 0 0 0 2.73.73l.15-.08a2 2 0 0 1 2 0l.43.25a2 2 0 0 1 1 1.73V20a2 2 0 0 0 2 2h.44a2 2 0 0 0 2-2v-.18a2 2 0 0 1 1-1.73l.43-.25a2 2 0 0 1 2 0l.15.08a2 2 0 0 0 2.73-.73l.22-.38a2 2 0 0 0-.73-2.73l-.15-.08a2 2 0 0 1 0-2l.15-.08a2 2 0 0 0 .73-2.73l-.22-.38a2 2 0 0 0-2.73-.73l-.15.08a2 2 0 0 1-2 0l-.43-.25a2 2 0 0 1-1-1.73V4a2 2 0 0 0-2-2z"></path>
-    <circle cx="12" cy="12" r="3"></circle>
-  </svg>
-);
-
-type ToggleSwitchProps = { isToggled: boolean; onToggle: () => void };
-const ToggleSwitch = ({ isToggled, onToggle }: ToggleSwitchProps) => (
-  <button
-    onClick={onToggle}
-    className={`relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none ${
-      isToggled ? "bg-primary" : "bg-gray-600"
-    }`}
-  >
-    <span
-      className={`pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out ${
-        isToggled ? "translate-x-5" : "translate-x-0"
-      }`}
-    />
-  </button>
-);
-
-const durationSteps = [
-  15, 30, 45, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 360,
-];
-
-const formatDuration = (minutes: number) => {
-  if (minutes < 60) {
-    return `${minutes} minutes`;
-  }
-  const hours = minutes / 60;
-  return `${hours} hour${hours > 1 ? "s" : ""}`;
-};
-
-export default function App() {
-  const [user, setUser] = useState<User | null>(null);
-  const [date, setDate] = useState<Date | undefined>(new Date());
-  type Task = {
-    id: string;
-    title: string;
-    is_time_sensitive: boolean;
-    duration_minutes: number | null;
-    scheduled_time: string | null;
-    is_completed: boolean;
-  };
-  type Settings = {
-    timetable_start: number;
-    timetable_end: number;
-    show_time_needle: boolean;
-  };
-
-  const [tasks, setTasks] = useState<Task[]>([]);
-  const [settings, setSettings] = useState<Settings | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [tasksLoading, setTasksLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const [newTaskTitle, setNewTaskTitle] = useState("");
-  const [isTimeSensitive, setIsTimeSensitive] = useState(false);
-  const [duration, setDuration] = useState(30);
-  const [scheduledTime, setScheduledTime] = useState("");
-  const [isAddingTask, setIsAddingTask] = useState(false);
-
-  const [reschedulingTask, setReschedulingTask] = useState<Task | null>(null);
-  const [isSettingsOpen, setIsSettingsOpen] = useState(false);
-
-  const router = useRouter();
-  const supabase = createBrowserClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  );
-
-  const fetchTasks = useCallback(async (selectedDate: Date) => {
-    setTasksLoading(true);
-    try {
-      const year = selectedDate.getFullYear();
-      const month = (selectedDate.getMonth() + 1).toString().padStart(2, "0");
-      const day = selectedDate.getDate().toString().padStart(2, "0");
-      const dateString = `${year}-${month}-${day}`;
-      const response = await fetch(`/api/tasks?date=${dateString}`);
-      if (!response.ok) {
-        throw new Error("Failed to fetch tasks");
-      }
-      const fetchedTasks = await response.json();
-      setTasks(fetchedTasks);
-    } catch (err: unknown) {
-      setError(
-        err instanceof Error ? err.message : "An unknown error occurred"
-      );
-    } finally {
-      setTasksLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    const checkUserAndFetch = async () => {
-      setLoading(true);
-      const {
-        data: { session },
-      } = await supabase.auth.getSession();
-      if (!session) {
-        router.push("/login");
-        return;
-      }
-      setUser(session.user);
-
-      try {
-        const settingsResponse = await fetch("/api/settings");
-        if (!settingsResponse.ok) {
-          throw new Error("Failed to fetch settings");
+// Client-only redirect component
+import dynamic from "next/dynamic";
+const ClientRedirect = dynamic(
+  () =>
+    Promise.resolve(function ClientRedirect() {
+      React.useEffect(() => {
+        function getCookie(name: string) {
+          if (typeof document === "undefined") return null;
+          const value = `; ${document.cookie}`;
+          const parts = value.split(`; ${name}=`);
+          if (parts.length === 2)
+            return parts.pop()?.split(";").shift() || null;
+          return null;
         }
-        const fetchedSettings = await settingsResponse.json();
-        setSettings(fetchedSettings);
-      } catch (err: unknown) {
-        setError(
-          err instanceof Error ? err.message : "An unknown error occurred"
-        );
-      }
-      setLoading(false);
-    };
-    checkUserAndFetch();
-  }, [router, supabase.auth]);
+        const token = getCookie(SESSION_COOKIE);
+        if (token) {
+          window.location.href = SCHEDULE_PATH;
+        }
+      }, []);
+      return null;
+    }),
+  { ssr: false }
+);
 
-  useEffect(() => {
-    if (date && user) {
-      fetchTasks(date);
-    }
-  }, [date, fetchTasks, user]);
-
-  const toLocalISOString = (dateToFormat: Date) => {
-    const year = dateToFormat.getFullYear();
-    const month = (dateToFormat.getMonth() + 1).toString().padStart(2, "0");
-    const day = dateToFormat.getDate().toString().padStart(2, "0");
-    return `${year}-${month}-${day}`;
-  };
-
-  const handleAddTask = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!newTaskTitle.trim() || isAddingTask) return;
-    setIsAddingTask(true);
-
-    const isForToday = date
-      ? new Date().toDateString() === date.toDateString()
-      : true;
-    const now = new Date();
-    const pad = (n: number) => n.toString().padStart(2, "0");
-    const localTimeString = `${now.getFullYear()}-${pad(
-      now.getMonth() + 1
-    )}-${pad(now.getDate())}T${pad(now.getHours())}:${pad(
-      now.getMinutes()
-    )}:${pad(now.getSeconds())}`;
-    let finalScheduledTime = null;
-    if (isTimeSensitive && scheduledTime && date) {
-      const [hours, minutes] = scheduledTime.split(":").map(Number);
-      const combinedDate = new Date(date);
-      combinedDate.setHours(hours, minutes, 0, 0);
-      finalScheduledTime = combinedDate.toISOString();
-    }
-
-    try {
-      const response = await fetch("/api/tasks", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          title: newTaskTitle,
-          is_time_sensitive: isTimeSensitive,
-          duration_minutes: Number(duration),
-          scheduled_time: finalScheduledTime,
-          user_current_time: localTimeString,
-          for_date: toLocalISOString(date || new Date()),
-          is_for_today: isForToday,
-          timezone_offset: new Date().getTimezoneOffset(),
-        }),
-      });
-      if (!response.ok) throw new Error("Failed to add task");
-      const addedTask = await response.json();
-      setTasks((prev) => [...prev, addedTask]);
-      setNewTaskTitle("");
-      setIsTimeSensitive(false);
-      setDuration(30);
-      setScheduledTime("");
-    } catch (err) {
-      setError(
-        err instanceof Error ? err.message : "An unknown error occurred"
-      );
-    } finally {
-      setIsAddingTask(false);
-    }
-  };
-
-  const handleToggleDone = async (taskId: string, currentStatus: boolean) => {
-    try {
-      const response = await fetch("/api/tasks", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ id: taskId, is_completed: !currentStatus }),
-      });
-      if (!response.ok) throw new Error("Failed to update task status");
-      const updatedTask = await response.json();
-      setTasks(tasks.map((task) => (task.id === taskId ? updatedTask : task)));
-    } catch (err) {
-      setError(
-        err instanceof Error ? err.message : "An unknown error occurred"
-      );
-    }
-  };
-
-  const handleDeleteTask = async (taskId: string) => {
-    setTasks(tasks.filter((task) => task.id !== taskId));
-    try {
-      const response = await fetch("/api/tasks", {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ id: taskId }),
-      });
-      if (!response.ok) throw new Error("Failed to delete task");
-    } catch (err) {
-      setError(
-        err instanceof Error ? err.message : "An unknown error occurred"
-      );
-    }
-  };
-
-  const handleOpenRescheduleModal = (task: Task) => setReschedulingTask(task);
-  const handleCloseRescheduleModal = () => setReschedulingTask(null);
-
-  const handleUpdateTaskTime = async (taskId: string, newTime: string) => {
-    try {
-      const response = await fetch("/api/tasks", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ id: taskId, scheduled_time: newTime }),
-      });
-      if (!response.ok) throw new Error("Failed to reschedule task");
-      const updatedTask = await response.json();
-      setTasks(tasks.map((task) => (task.id === taskId ? updatedTask : task)));
-      handleCloseRescheduleModal();
-    } catch (err) {
-      setError(
-        err instanceof Error ? err.message : "An unknown error occurred"
-      );
-    }
-  };
-
-  const handleSaveSettings = async (newSettings: Settings) => {
-    try {
-      const response = await fetch("/api/settings", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(newSettings),
-      });
-      if (!response.ok) throw new Error("Failed to save settings");
-      const updatedSettings = await response.json();
-      setSettings(updatedSettings);
-      setIsSettingsOpen(false);
-    } catch (err) {
-      setError(
-        err instanceof Error ? err.message : "An unknown error occurred"
-      );
-    }
-  };
-
-  if (loading || !settings) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center text-foreground">
-        Loading...
-      </div>
-    );
-  }
-
+function Landing() {
   return (
-    <div className="min-h-screen bg-background text-foreground font-sans antialiased">
-      <header className="border-b border-border">
-        <div className="container mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex h-16 items-center justify-between">
-            <div className="flex items-center space-x-3">
-              <PlazenLogo />
-              <span className="text-xl font-semibold">Plazen</span>
-            </div>
-            <div className="flex items-center space-x-2">
-              <Button
-                variant="ghost"
-                size="icon"
-                onClick={() => setIsSettingsOpen(true)}
-              >
-                <SettingsIcon />
-              </Button>
-              <Button
-                variant="ghost"
-                size="icon"
-                onClick={() => router.push("/account")}
-              >
-                <UserIcon />
-              </Button>
-            </div>
-          </div>
-        </div>
-      </header>
-
-      <main className="container mx-auto px-4 sm:px-6 lg:px-8 py-8">
-        {error && (
-          <div className="bg-destructive/20 border border-destructive text-destructive-foreground rounded-md p-4 mb-6">
-            {error}
-          </div>
-        )}
-        <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
-          <div className="lg-col-span-1 space-y-8">
-            <div className="bg-card rounded-lg shadow-lg p-6 border border-border">
-              <h2 className="text-lg font-medium mb-4">Add a New Task</h2>
-              <form onSubmit={handleAddTask} className="space-y-4">
-                <div>
-                  <label htmlFor="task-title" className="sr-only">
-                    Task Title
-                  </label>
-                  <input
-                    type="text"
-                    id="task-title"
-                    value={newTaskTitle}
-                    onChange={(e) => setNewTaskTitle(e.target.value)}
-                    className="mt-1 block w-full rounded-md bg-input border-border text-foreground shadow-sm focus:border-primary focus:ring-primary sm:text-sm"
-                    placeholder="e.g., Water the plants"
-                  />
-                </div>
-                <div className="flex items-center justify-between">
-                  <span className="text-sm font-medium text-muted-foreground">
-                    Time Sensitive?
-                  </span>
-                  <ToggleSwitch
-                    isToggled={isTimeSensitive}
-                    onToggle={() => setIsTimeSensitive(!isTimeSensitive)}
-                  />
-                </div>
-
-                {isTimeSensitive && (
-                  <div>
-                    <label
-                      htmlFor="scheduled-time"
-                      className="block text-sm font-medium text-muted-foreground mb-1"
-                    >
-                      Start Time
-                    </label>
-                    <input
-                      type="time"
-                      id="scheduled-time"
-                      value={scheduledTime}
-                      onChange={(e) => setScheduledTime(e.target.value)}
-                      className="mt-1 block w-full rounded-md bg-input border-border text-foreground shadow-sm focus:border-primary focus:ring-primary sm:text-sm"
-                    />
-                  </div>
-                )}
-
-                <div className="space-y-2">
-                  <div className="flex justify-between items-center">
-                    <label
-                      htmlFor="duration"
-                      className="text-sm font-medium text-muted-foreground"
-                    >
-                      Duration
-                    </label>
-                    <span className="text-sm font-semibold text-foreground">
-                      {formatDuration(duration)}
-                    </span>
-                  </div>
-                  <Slider
-                    id="duration"
-                    min={0}
-                    max={durationSteps.length - 1}
-                    step={1}
-                    value={[durationSteps.indexOf(duration)]}
-                    onValueChange={(value) =>
-                      setDuration(durationSteps[value[0]])
-                    }
-                  />
-                </div>
-
-                <Button
-                  type="submit"
-                  className="w-full"
-                  disabled={isAddingTask}
+    <>
+      <ClientRedirect />
+      <main className={`${styles.page} ${styles.theme}`}>
+        <header className={styles.header}>
+          <div className={styles.container}>
+            <div className={styles.topBar}>
+              <div className={styles.brand}>
+                <span className={styles.brandMark} />
+                <span className={styles.brandText}>Plazen</span>
+              </div>
+              <nav className={styles.nav}>
+                <a href="#how" className={styles.navLink}>
+                  How it works
+                </a>
+                <a href="#features" className={styles.navLink}>
+                  Features
+                </a>
+                <a href="#faq" className={styles.navLink}>
+                  FAQ
+                </a>
+                <Link
+                  href="/login"
+                  className={`${styles.navLink} ${styles.loginLink}`}
                 >
-                  {isAddingTask ? (
-                    "Adding..."
-                  ) : (
-                    <>
-                      <PlusIcon className="mr-2" />
-                      Add to Schedule
-                    </>
-                  )}
-                </Button>
-              </form>
-            </div>
-            <div className="bg-card rounded-lg shadow-lg border border-border">
-              <Calendar
-                mode="single"
-                selected={date}
-                onSelect={setDate}
-                className="p-4 w-full"
-              />
+                  Log in
+                </Link>
+              </nav>
             </div>
           </div>
-
-          <div className="lg:col-span-2">
-            <Timetable
-              tasks={tasks}
-              settings={settings}
-              date={date || new Date()}
-              tasksLoading={tasksLoading}
-              onToggleDone={handleToggleDone}
-              onDeleteTask={handleDeleteTask}
-              onReschedule={handleOpenRescheduleModal}
+        </header>
+        <section className={styles.hero}>
+          <div className={styles.heroText}>
+            <h1 className={styles.title}>Your day, planned for you.</h1>
+            <p className={styles.subtitle}>
+              Plazen turns tasks into time. Add flexible to‑dos, pin hard
+              appointments, and let Plazen place everything into a realistic
+              daily schedule—no dragging blocks or micromanaging.
+            </p>
+            <div className={styles.ctaRow}>
+              <Link href="/login" className={styles.buttonPrimary}>
+                Get started — Log in
+              </Link>
+              <a href="#how" className={styles.buttonSecondary}>
+                See how it works
+              </a>
+            </div>
+            <p className={styles.trustNote}>
+              Built for people who value deep work: students, founders,
+              creators, and focused teams.
+            </p>
+          </div>
+          <div className={styles.heroMedia}>
+            <Image
+              width={800}
+              height={600}
+              className={styles.heroImage}
+              src="/og-image.png"
+              alt="Plazen daily plan with tasks placed between calendar events"
+              loading="eager"
+            />
+            <Image
+              width={800}
+              height={600}
+              className={styles.mediaImage}
+              src="/images/mobile.png"
+              alt="Mobile view of Plazen with today's plan"
+              loading="lazy"
             />
           </div>
-        </div>
-      </main>
+        </section>
 
-      {reschedulingTask && (
-        <RescheduleModal
-          task={reschedulingTask}
-          onClose={handleCloseRescheduleModal}
-          onSave={handleUpdateTaskTime}
-        />
-      )}
-      {isSettingsOpen && (
-        <SettingsModal
-          currentSettings={settings}
-          onClose={() => setIsSettingsOpen(false)}
-          onSave={handleSaveSettings}
-        />
-      )}
-    </div>
+        <section id="features" className={styles.sectionAlt}>
+          <div className={styles.container}>
+            <h2 className={styles.sectionTitle}>
+              Built for focus, not fiddling
+            </h2>
+            <ul className={styles.featureList}>
+              <li className={styles.featureItem}>
+                <span className={styles.featureName}>Focus mode</span>
+                <span className={styles.featureDesc}>
+                  One task at a time with gentle, timely handoffs.
+                </span>
+              </li>
+              <li className={styles.featureItem}>
+                <span className={styles.featureName}>Smart durations</span>
+                <span className={styles.featureDesc}>
+                  Learns how long tasks really take to improve future plans.
+                </span>
+              </li>
+              <li className={styles.featureItem}>
+                <span className={styles.featureName}>Energy‑aware</span>
+                <span className={styles.featureDesc}>
+                  Schedule deep work when you’re sharp; save admin for later.
+                </span>
+              </li>
+              <li className={styles.featureItem}>
+                <span className={styles.featureName}>Calendar‑friendly</span>
+                <span className={styles.featureDesc}>
+                  Respects your events so plans reflect reality.
+                </span>
+              </li>
+              <li className={styles.featureItem}>
+                <span className={styles.featureName}>Deadline protection</span>
+                <span className={styles.featureDesc}>
+                  Prioritizes time‑sensitive items so nothing slips.
+                </span>
+              </li>
+              <li className={styles.featureItem}>
+                <span className={styles.featureName}>Tweakable</span>
+                <span className={styles.featureDesc}>
+                  Nudge a task; Plazen rebalances the rest intelligently.
+                </span>
+              </li>
+            </ul>
+            <Image
+              width={800}
+              height={600}
+              className={styles.integrationsImage}
+              src="/images/integrations.png"
+              alt="Integrations and calendar syncing"
+              loading="lazy"
+            />
+          </div>
+        </section>
+
+        <section className={styles.section}>
+          <div className={styles.container}>
+            <h2 className={styles.sectionTitle}>What people say</h2>
+            <div className={styles.testimonialGrid}>
+              <figure className={styles.testimonial}>
+                <blockquote className={styles.testimonialQuote}>
+                  “I stopped spending 30 minutes every morning shuffling tasks.
+                  Plazen just handles it, so I actually start working.”
+                </blockquote>
+                <figcaption className={styles.testimonialMeta}>
+                  <Image
+                    width={800}
+                    height={600}
+                    className={styles.testimonialAvatar}
+                    src="/images/testimonial-1.jpg"
+                    alt="Samira A."
+                  />
+                  <div className={styles.testimonialIdentity}>
+                    <span className={styles.testimonialName}>Samira A.</span>
+                    <span className={styles.testimonialRole}>
+                      Founder & Researcher
+                    </span>
+                  </div>
+                </figcaption>
+              </figure>
+
+              <figure className={styles.testimonial}>
+                <blockquote className={styles.testimonialQuote}>
+                  “My days finally feel realistic. I finish more and worry less.
+                  It’s like a calm project manager in my pocket.”
+                </blockquote>
+                <figcaption className={styles.testimonialMeta}>
+                  <Image
+                    width={800}
+                    height={600}
+                    className={styles.testimonialAvatar}
+                    src="/images/testimonial-2.jpg"
+                    alt="Jon P."
+                  />
+                  <div className={styles.testimonialIdentity}>
+                    <span className={styles.testimonialName}>Jon P.</span>
+                    <span className={styles.testimonialRole}>
+                      Product Designer
+                    </span>
+                  </div>
+                </figcaption>
+              </figure>
+
+              <figure className={styles.testimonial}>
+                <blockquote className={styles.testimonialQuote}>
+                  “Pin the must‑dos, let Plazen place the rest. It’s the first
+                  task manager that respects my time.”
+                </blockquote>
+                <figcaption className={styles.testimonialMeta}>
+                  <Image
+                    width={800}
+                    height={600}
+                    className={styles.testimonialAvatar}
+                    src="/images/testimonial-3.jpg"
+                    alt="Nina C."
+                  />
+                  <div className={styles.testimonialIdentity}>
+                    <span className={styles.testimonialName}>Nina C.</span>
+                    <span className={styles.testimonialRole}>Grad Student</span>
+                  </div>
+                </figcaption>
+              </figure>
+            </div>
+          </div>
+        </section>
+
+        <section id="faq" className={styles.sectionAlt}>
+          <div className={styles.container}>
+            <h2 className={styles.sectionTitle}>Frequently asked questions</h2>
+
+            <details className={styles.faqItem}>
+              <summary className={styles.faqSummary}>
+                How is Plazen different from a normal to‑do app?
+              </summary>
+              <p className={styles.faqText}>
+                Most to‑do apps collect tasks. Plazen turns tasks into time. It
+                plans your day automatically, balancing fixed events with
+                flexible work so you don’t have to micromanage your schedule.
+              </p>
+            </details>
+
+            <details className={styles.faqItem}>
+              <summary className={styles.faqSummary}>
+                Can I still drag and drop tasks?
+              </summary>
+              <p className={styles.faqText}>
+                Yes. You can nudge and reorder. Plazen will reflow everything
+                else intelligently, preserving your anchors.
+              </p>
+            </details>
+
+            <details className={styles.faqItem}>
+              <summary className={styles.faqSummary}>
+                Does it work with my calendar?
+              </summary>
+              <p className={styles.faqText}>
+                Plazen respects your calendar events and can sync with popular
+                providers so your schedule stays true to life.
+              </p>
+            </details>
+
+            <details className={styles.faqItem}>
+              <summary className={styles.faqSummary}>
+                Is there a free plan?
+              </summary>
+              <p className={styles.faqText}>
+                Yes. Get started for free and upgrade if you need more power.
+              </p>
+            </details>
+          </div>
+        </section>
+
+        <section className={styles.ctaSection}>
+          <div className={styles.container}>
+            <h2 className={styles.ctaTitle}>Ready to reclaim your day?</h2>
+            <p className={styles.ctaText}>
+              Join people who plan less and accomplish more with Plazen.
+            </p>
+            <Link href="/login" className={styles.buttonPrimaryLarge}>
+              Log in to start planning
+            </Link>
+            <p className={styles.smallNote}>No credit card required.</p>
+          </div>
+        </section>
+
+        <footer className={styles.footer}>
+          <div className={`${styles.container} ${styles.footerInner}`}>
+            <div className={styles.brand}>
+              <span className={styles.brandMark} />
+              <span className={styles.brandText}>Plazen</span>
+            </div>
+            <nav className={styles.footerNav}>
+              <a href="#features" className={styles.footerLink}>
+                Features
+              </a>
+              <a href="#how" className={styles.footerLink}>
+                How it works
+              </a>
+              <a href="#faq" className={styles.footerLink}>
+                FAQ
+              </a>
+              <Link href="/login" className={styles.footerLink}>
+                Log in
+              </Link>
+            </nav>
+            <p className={styles.copyright}>
+              © {new Date().getFullYear()} Plazen
+            </p>
+          </div>
+        </footer>
+      </main>
+    </>
   );
 }
